fixing logic for add calculation

+++ src/main/java/com/irfaan/taxcalculation/strategy/TaxStrategy.java
@@ -2,6 +2,6 @@ package com.irfaan.taxcalculation.strategy;
 
 public interface TaxStrategy {
 
-    long calculateBasedNationality(String gender, long incomePerMonth, String maritalStatus, long nonDeductibleEarning);
+    double calculateBasedNationality(String gender, double incomePerMonth, String maritalStatus, double nonDeductibleEarning);
 
 }
